Screen Header - Refatoração e Flexbox
