+++ src/components/Screen/components/ScreenHeader.tsx
@@ -1,0 +1,36 @@
+import React from 'react';
+
+import {useNavigation} from '@react-navigation/native';
+
+import {Box, Icon, Text, TouchableOpacityBox} from '@components';
+
+import {ScreenProps} from '../Screen';
+
+const ICON_SIZE = 20;
+type Props = Pick<ScreenProps, 'title' | 'canGoBack'>;
+export function ScreenHeader({title, canGoBack}: Props) {
+  const navigation = useNavigation();
+  return (
+    <Box
+      flexDirection="row"
+      alignItems="center"
+      justifyContent="space-between"
+      mb="s24">
+      {canGoBack && (
+        <TouchableOpacityBox
+          onPress={navigation.goBack}
+          flexDirection="row"
+          alignItems="center">
+          <Icon size={ICON_SIZE} name="arrowLeft" color="primary" />
+          {!title && (
+            <Text preset="paragraphMedium" semibold ml="s8">
+              Voltar
+            </Text>
+          )}
+        </TouchableOpacityBox>
+      )}
+      {title && <Text preset="headingSmall">{title}</Text>}
+      {title && <Box width={ICON_SIZE} />}
+    </Box>
+  );
+}
